packages/vision-service/requirements.txt
# Vision Service Dependencies
# Using newer versions compatible with Python 3.14
opencv-python>=4.9.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
